archive: update C struct definitions and types of fields
Update C struct definitions in comments of archive writer/reader
code.

Update the `struct.pack`/`struct.unpack` format strings to match
the signedness of fields. I.e., update format code from `i` to `I`
for fields whose type in C code was updated from `int` to `uint32_t`.
Also use `c` format code instead of `B` for the `typecode` character
in the  CArchive TOC, as we are treating the field as `char` and
not an `unsigned char`.