bootloader: create copy of argv only if needed
Create a copy of command line arguments only if we really need it;
i.e., in windowed bootloader for macOS .app bundles (where we need
to filter out `-psn_xxx` argument, and may append additional
arguments if argv-emulation is enabled).

For regular POSIX codepaths, we can spawn onefile child process with
`execvp` that is called with original `argv`, because the C standard
requirements guarantee us that `argv[argc] = NULL`, and thus we do not
need to copy arguments into a NULL-terminated array.

This trivializes the implementation of `pyi_utils_replace_process`
into a simple `execvp` call, which we can inline and remove the
helper.