bootloader: clean up win32 character set conversion helpers
Clean up the low-level win32 character set conversion helpers (and
the winerror code-to-string conversion function).

Rename the functions:
- `GetWinErrorString` -> `pyi_win32_get_winerror_string`
- `pyi_win32_utils_from_utf8` -> `pyi_win32_utf8_to_wcs`
- `pyi_win32_utils_to_utf8` -> `pyi_win32_wcs_to_utf8`

Swap the order of source and destination arguments so that source
comes before destination.

Remove calls to FATAL_WINERROR in error codepaths of these helpers;
these are low-level helpers, and should not be popping the error
dialog. Moreso because they are also used in the said error dialog
codepath...