tests: qt: skip test_Qt_QtQml with PySide6 and PyQt6 6.6.3
Qt6 6.6.3 split Qt Quick Controls 2 styles into separate shared
libraries, and both `PySide6` 6.6.3 and `PyQt6` 6.6.3 PyPI wheels
failed to account for that.

The shared libs are missing from `PySide6-Essentials` 6.6.3 on
all OSes. The problem was fixed by subsequent 6.6.3.1 and 6.7.0
releases.

The shared libs are missing from `PyQt6-Qt6` 6.6.3 on Windows and
Linux, but are present on macOS.

So skip the corresponding variant of `test_Qt_QtQml` if running
with affected version of bindings.