depend: fix library path resolution in POSIX non-ldconfig branches
The `resolve_library_path` helper in `PyInstaller.depend.bindepend`
expects to be given a full library name; on POSIX systems, this
includes `lib` prefix and `.so` suffix - typically versioned,
as it appears in the soname.

This means that in `_resolve_library_path_unix` codepaths that do
not use `ldconfig` cache but rather search for the library in
given search locations, we should not be using `_which_library`
helper (which expects basename without suffix), but rather
`_resolve_library_path_in_search_paths` helper (which expects
full library name).

However, because it was previously possible to pass a basename
without suffix and because some hooks in contrib hooks repository
rely on this behavior, implement support for both; if the given
name has .so suffix, use  `_resolve_library_path_in_search_paths`
to find exact match, and if name has no suffix, use `_which_library`
to find matching basename instead.