depend: do not resolve soname in library path resolution
In `_resolve_library_path_unix` helper, remove the part at the
end where we read the soname of the resolved library and replace
the name in returned path with the soname.

In most cases, this was effectively no-op, because we use this
function to resolve dependencies of binaries that we analyze,
and those refer to those dependencies via their sonames.

However, in the unlikely case that we are trying to resolve a
library with soname that differs from its true (file)name, we
would need to collect it under that true (file)name; first, because
whoever was referring to it expects to find it under the true
(file)name, and second, becase in such case, the file with
corresponding soname might not actually exist at all.