building: do not attempt to expand env. vars in paths
Remove the `compat.expand_path` helper, and stop attempting to
expand environment variables in paths given via `--workpath`,
`--distpath`, `--specpath`, and `--additional-hooks-dir`.

The Windows implementation of `expandvars` found in `ntpath`
translates `$$` and `%%` into `$` and `%`, respectively, which
mangles the paths that contain such sequences (as literal characters,
without using environment variables at all).

While the above issue could be worked around by duplicating characters
in the problematic character sequences, the whole environment
variable expansion in paths seems rather unnecessary. It was introduced
in response to #696, which was caused by the fact that shell does
not expand the tilde when using `--argname=~/path/abc` argument syntax
instead of `--argname ~/path/abc` (or when the path argument is
quoted).

As that seems like legitimate shortcoming, keep the tilde expansion
part (`os.path.expanduser`) around, but remove the general environment
variable expansion attempts (`os.path.expandvars`), which should
be left to the shell.