doc: Encourage keeping pyinstaller and hooks-contrib version in sync
Users who are naive to pyinstaller-hooks-contrib's existence may pin
pyinstaller but not the hooks repo for long periods of time, meaning
that the pyinstaller-hooks-contrib installations can end up with
pyinstaller versions that lack the hook utilities it depends on.

This can ultimately lead us to having to write hooks that are
indefinitely compatible with ``pyinstaller==4.0`` which we don't want to
do so at least reduce the frequency of this by trying to raise awareness
of the hooks package.

See pyinstaller/pyinstaller-hooks-contrib#736