bootloader: Windows: additional verification of runtime-tmpdir
Implement additional verification of `runtime-tmpdir` path on
Windows. If the path is just a drive root, ensure that the drive
exists (i.e., is recongized by `GetDriveTypeW`). Otherwise, ensure
that when attempting to create the resolved absolute path, the
very last `CreateDirectoryW` call either succeeds or fails with
`ERROR_ALREADY_EXISTS`.

These checks aim to prevent us from feeding invalid path to
`_wtempnam` later on, which would cause it to return `\MEIXXXXXX`
("using current working directory" as MSDN puts it), which ends up
being erroneously resolved as `c:\MEIXXXXXX˙ (or whatever the
current drive is).