tests: importlib.resources: update check for functional API
It looks like the re-introduced functional API from
importlib-resources 6.4.0 was also back-ported to python stdlib
in 3.13b1, but the sub-module that we use to detect presence of
this API was renamed from `functional` to `_functional`. Check
for both names, as importlib-resources also seems to be planning
to rename that module.