tests: isolated: adjust expected file handles increase for python 3.13
It looks like that with python 3.13, the increase in number of
opened file handles in `test_pipe_leakage` on Windows is down from
six to three.

Spawning the child process now seems to open only one handle instead
of two.

Opening python file objects on top of the file descriptors (that are
in turn opened on top of the two pipe handles) does not seem to
open any additional handles anymore.