hook: win32ctypes.core: always collect the ctypes backend
Always collect the `ctypes` backend (`win32ctypes.core.ctypes` modules),
even if we are also collecting the `cffi` backend due to `cffi`
being available in the build environment.

This prevents issues when import of `cffi` is blocked at run-time
(similarly to how we do it to avoid issues with -OO optimization
level) or when `cffi` is excluded from collection by user (for
example, by using `--exclude-module cffi`).