building: ensure symbolic links on Windows use backslash separators
On Windows, ensure that created symbolic links contain backslash
as path component separator. In order to handle all cases, this
enforcement is performed at the lowest level possible - i.e., in
`COLLECT` when creating symbolic link for onedir build, or in
`CArchiveWriter` when writing PKG TOC entry for onefile build.