bootloader: Windows: ensure stderr is unbuffered regardless of compiler
Ensure that stderr is unbuffered on Windows regardless of compiler.
Up until now, we did that only for MSVC, but it looks like both
Visual C runtime and MinGW end up buffering stderr when it is
redirected.

This explains why the splash screen tests failed to produce any
output to stderr when running in msys/mingw environment (and
subsequently failed due to not being able to find a specific string
in the said output). In contrast to most tests where the frozen
 application automatically exits (and timeout is set to detect
failures such as endless loop), in this casem the application is
does not exit automatically; we keep it running for 10 seconds (so
reaching timeout indicates success), then we kill it. And if stderr
is buffered, its contents are lost.