hookutils: qt: suppress warnings for unavailable Qt bindings packages
When trying to obtain Qt information as part of the delayed
initialization of `QtLibraryInfo`, explicitly handle `ModuleNotFound`
error in order to avoid raising a warning message for bindings package
that is unavailable.

The delayed initialization might be triggered by a helper function
(e.g., the `exclude_extraneous_qt_bindings` helper used by `qtpy`
hook and other hooks for Qt-based packages) trying to determine
availability of bindings by querying the `version` attribute of
the `QtLibraryInfo` instances for all bindings.