Add support for python builds with PEP703 free-threading enabled
Implement support for python builts with PEP703 free-threading enabled,
i.e., built with `--disable-gil` option (`Py_GIL_DISABLED`).

On build side, this involves detecting such a build, which is done via
`sysconfig.get_config_var('Py_GIL_DISABLED')`. Add an equivalent
`compat.is_nogil` flag to cache the obtained value.

In free-threading-enabled build, the python shared library has "t"
ABI suffix, so adjust `compat.PYDYLIB_NAMES` accordingly. We strictly
enforce this suffix (or the lack thereof) to prevent accidental mix-up
of shared libraries when multiple different builds can be discovered
in the library search path.

If free-threading is enabled, we add `pyi-python-flag Py_GIL_DISABLED`
run-time OPTION to PKG archive, and have bootloader parse this
option to recognize that free-threading is enabled.

On bootloader side, we need to keep two variants of `PyConfig` structure
for python 3.13; one with `enable_gil` field (`Py_GIL_DISABLED defined`)
and one without it (`Py_GIL_DISABLED` not defined). Unfortunately, this
field was added in the middle of the config structure, and we need to
access fields that come after it, so the layout matters. (Although
strictly speaking, this applies only to 32-bit builds; on 64-bit
plaftorms, presence/absence of this field makes no difference to the
layout of subsequent fields, due to structure alignment rules).

Based on presence or absence of afore mentioned OPTION passed from our
build process, the bootloader selects the correct `PyConfig` layout
to use. We do no need to interact with the `enable_gil` field; if
user wants to force-enable or force-disable GIL (the equivalent of
python's `-X gil=0` and `-X gil=1`, or `PYTHON_GIL` environment
variable), they can do so via `--python-option`, which already
supports pass-through for X-options. For the sake of completeness,
add an example to Specifying Python  Interpreter Options documentation
section.