tests: add test for environment inheritance by spawned subprocess
Add a test that explicitly checks for environment inheritance by
spawned subprocess. Specifically, we are interested in propagation
of top-level application directory (`sys._MEIPASS`) in following
scenarios:
 - onedir application spawns subprocess using `sys.executable`
 - onefile application spawns subprocess using `sys.executable`
 - onedir application spawns subprocess with another onedir executable
 - onedir application spawns subprocess with onefile executable
 - onefile application spawns subprocess with another onefile executable
 - onefile application spawns subprocess with onedir executable

In the first two scenarios, the subprocess should end up using the
same `sys._MEIPASS` as the parent process (in the case of onefile
application, the executable should not unpack itself again).

In the remaining four scenarios, the child process should end up
with a different `sys._MEIPASS` (and when child process is a
onefile executable, it should unpack itself).

Currently, this behavior is achieved by bootloader resetting
`_MEIPASS2` environment variable before running the program code.
This caters to the last four scenarios (i.e., application implicitly
assumes that it is running in a different environment). The first two
cases require the user to restore `_MEIPASS2` environment variable,
as currently done in the test program as well (and, for example,
in `multiprocessing` run-time hook).

Note that this is far from ideal, because it requires all users
of `sys.executable` to explicitly restore `_MEIPASS2` environment
variable before spawning sub-process. The plan is to rework the
bootloader logic to automatically handle all of above scenarios,
without user's intervention, and the added test will serve for
validation.