bootloader: rework the detection of inherited PyInstaller environment
Rework the logic for detection of existing PyInstaller environment
to assume inherited environment by default instead of assuming
non-inherited (isolated) environment.

This logic is primarily needed for onefile applications. The existing
implementation went as follows: the parent process of a onefile application
unpacks the archive into ephemeral application's top-level directory,
and sets the `_MEIPASS2` environment variable to pass this path on to
the child process. The child process, after setting up its paths,
clears this environment variable before running the user's code.

This clearing of `_MEIPASS2` is done in case the user's code tries
to spawn a sub-process using a different onefile application; in such
case, the presence of `_MEIPASS2` environment variable would "trick"
this second application process into thinking that it is a child process
of an already-unpacked application, and re-using the ephemeral
application top-level directory of the first application.

On the other hand, this means that if we want to spawn a sub-process
using the same executable (i.e., using `sys.executable`) and re-use
the unpacked contents, we need to explicitly restore `_MEIPASS2`
environment variable prior to launching the subprocess.

While we do such restoring of `_MEIPASS2` in our `multiprocessing`
hook, this approach does not scale well. We would need to somehow
ensure that any use of `sys.executable` has `_MEIPASS2` restored; this
includes, but is not limited to, different alternative multi-processing
frameworks, `gooey`, manual spawning of sub-processes (either via
`subprocess` or external frameworks, such as Qt's `QProcess`).

So we are inverting the logic - if we find an existing PyInstaller
environment variables, we assume that we are inheriting that environment,
unless there is compelling evidence to the contrary. For this purpose,
we introduce new environment variable, `_PYI_ARCHIVE_FILE`, in which we
track the path to the PKG/CArchive file (usually the executable file
itself). If the path to archive file of the current process is
different from that stored in `_PYI_ARCHIVE_FILE`, we clear all
PyInstaller-related environment variables to initialize a clear
environment. Otherwise, we inherit the existing environment.

We also rename `_MEIPASS2` environment variable to
`_PYI_APPLICATION_HOME_DIR`, in order to avoid issues when spawning
executables made with older PyInstaller versions (which rely on the
old behavior of `_MEIPASS2` not being defined).