bootloader: wscript: add -mwindows only to LINKFLAGS
The `-mwindows` flag is meant for linker, so it should be specified
only under `LINKFLAGS`. While gcc does not seem to take issue
with it also being specified under `CFLAGS`, clang complains about
it; either in form of `-Wunused-command-line-argument` warnings,
or even errors (this seems applicable to recent versions of clang
coupled with MSVC linker).