docs: add a section on new requirement for implementing application restart
Add a new section to Common Issues and Pitfalls, which describes
the new requirement for spawning `sys.executable` subprocesses
that are supposed to outlive the current application process,
for example, to implement an application restart.

Due to changes made in previous commits, `_PYI_ARCHIVE_FILE` must
be now unset prior to spawning the subprocess in order for bootloader
to treat it as a new top-level application process and, for example,
have the onefile application unpack itself again.