Fix deprecated DEFAULT_FILE_STORAGE can be None.
settings.DEFAULT_FILE_STORAGE is deprecated since Django 4.2 and will be
removed in a future version (greater than 5.0).
In Django 5.0, not explicitely setting it to a value is equivalent to setting
it to None instead of the default value (undocummented behavior).
See https://docs.djangoproject.com/en/5.0/ref/settings/#default-file-storage