bootloader: debug macros: capture errno / GetLastError() at macro invocation
Capture the current value of errno or the value returned by
GetLastError() (on Windows) at the moment where debug/error macro
is invoked, instead of inside of the underlying formatting
function.

This minimizes the chances that the value of the error code changes
due to function calls made within the formatting functions themselves.

NOTE: some of the formatting functions are not completely ported
to this new behavior yet, as their implementations currently still
call `perror()` instead of `strerror()`.