bootloader: fix Windows console handler for Windows terminal
When console handler receives `CTRL_CLOSE_EVENT`, `CTRL_SHUTDOWN_EVENT`,
or `CTRL_LOGOFF_EVENT`, have it set a flag in `PYI_CONTEXT` before
going to sleep (to delay the OS terminating the process for as long
as possible). Have the main thread of onefile parent periodically
check the said flag; if the flag has been set, terminate the child
process (after a short grace period), and proceed with the cleanup
and shutdown (especially temporary directory cleanup).

This is necessary when Windows Terminal is used instead of old
`conhost.exe`; with the latter, the child process also received the
`CTRL_CLOSE_EVENT`, which terminated it. Therefore, parent process
detected that child is not running anymore, and proceeded with its
cleanup and shutdown. With Windows Terminal, this does not seem
to be the case anymore; the child does not receive the event until
after the OS has terminated the parent process (which was waiting
for the child to exit or be terminated). Thus explicit child
process termination is required, which is done after grace period
of 500 milliseconds.