bootloader: implement proper handling of Windows session shutdown in onefile mode
In onefile mode, have the parent process create a hidden window
to receive and handle `WM_QUERYENDSESSION` and `WM_ENDSESSION`
messages, which are sent by the OS when session shutdown is imminent
(i.e., user either requests to log off, or to perform system shutdown
or restart).

Properly handling these messages allows us to clean up the temporary
files before the OS ultimately terminates the process. This fixes
the leak of temporary directories of onefile applications that run
in the background (and thus need to be properly shut down when the
OS session is shutting down).

In console-enabled builds, this used to be handled via
`CTRL_LOGOFF_EVENT` and `CTRL_LOGOFF_EVENT` console events that
were processed by console event handler (which was introduced in
PyInstaller 5.3). However, it was subsequently broken starting
with PyInstaller 6.0 due to bootloader being linked against
`user32.dll` (for a similar reason, it never worked in splash-screen
enabled builds). In noconsole/windowed builds, it never worked due
to lack of console (and thus, the lack of console event handler).

The hidden-window approach should now work in both console
and windowed/noconsole builds, and regardless of whether splash
screen is used or not.

The messages for hidden window are processed in the main waiting
loop (where we also monitor child process and status flag controlled
by the console event handler), as well in subsequent steps after
exiting the main waiting loop. This is necessary to handle
different situations during session shutdown:
- the child process is terminated by the OS for not processing its
  copy of `WM_QUERYENDSESSION` message, and this happens before we
  receive `WM_QUERYENDSESSION`.
- the child process is terminated by the OS after we receive and
  process `WM_QUERYENDSESSION`, but before we receive `WM_ENDSESSION`.
- we receive and process `WM_QUERYENDSESSION` and `WM_ENDSESSION`
  while the child process is still alive (for example, because it
  also handles the messages and is delaying the shutdown).

In the first scenario, we cannot know whether the child stopped
due to `WM_QUERYENDSESSION` or for some other reason; to reliably
handle this situation, we now perform a short wait (250 ms) after
exiting the main wait loop, to see if we happen to receive
`WM_QUERYENDSESSION`. This introduces additional delay in normal
program shutdown situations, but is necessary to ensure that
if session shutdown is indeed imminent, we properly handle
`WM_QUERYENDSESSION` (and subsequent `WM_ENDSESSION`), instead of
proceeding with regular cleanup procedure, which might be
interrupted by the OS terminating the process for not processing
`WM_QUERYENDSESSION`.

If the child is still running when we enter `WM_ENDSESSION`
handler (the third scenario), we terminate it after grace period
of 1 second. In these the-world-is-about-to-end situations, not
leaking the temporary files takes precedence over event handling
logic that might be implemented in user's code.