bootloader: wscript: enable extra warnings when building with clang
When building with `clang`, enable extra warnings and treat them as
errors; i.e., add `-Wall` and `-Werror` to CFLAGS, same as we do
for `gcc`.

This flushes out the fact that we incorrectly check for availability
of `-Wl,--as-needed`; first, the flag is a linker flag and not
a compiler flag (i.e., the check should pass it via `linkflags`
instead of `cflags`); second, it should be marked as non-mandatory,
otherwise the whole check is pointless.