bootloader: archive: find SPLASH TOC entry during initial TOC scan
At the end of `pyi_archive_open`, we go through all TOC entries to
determine whether any of them are extractable (which indicates an
application with onefile semantics). We can reuse the same scan
to check whether a SPLASH entry is present, and store the pointer
to it.

This alleviates the need for another scan in `pyi_splash_setup`,
plus it decouples the ability to check for presence of splash
screen resources from loading them. Therefore, the
`_pyi_main_setup_splash_screen` helper can now exit early if
no splash screen is available.