bootloader: work around another source of VCRUNTIME140.dll leak
Have the parent process of a splash-screen-enabled onefile
application prefer the system-wide copy of `VCRUNTIME140.dll`
(if available), by trying to preemptively load it before splash
screen is set up.

Contemporary Tcl/Tk DLLs are known to depend on `VCRUNTIME140.dll`,
so when we load them during splash screen initialization, they
cause `VCRUNTIME140.dll` to be loaded as well. This happens after
the frozen application's top-level directory is pushed into the
search path; therefore, the bundled copy of the DLL is loaded, which
is intended to accommodate systems that do not have system-wide copy
of the DLL available.

However, the OS and/or anti-virus programs might inject additional
DLLs into the process that also depend on `VCRUNTIME140.dll`
(see the follow-up discussion under #7106); such externally loaded
DLLs might prevent us from unloading the bundled copy of
`VCRUNTIME140.dll` during the cleanup phase, and thus interfere with
removal of application's temporary directory.