loader: have PyiFrozenImporter.find_spec ignore non-string path entries
In `PyiFrozenImporter.find_spec`, skip entries of `path` that are not
instances of `str`. This matches behavior of built-in finders, which
perform similar entry filtering. In our case, skipping non-string
entries (either `None` or some other type) avoids raising an
`AttributeError` when trying to compare path prefix via `startswith`
method.