makespec: disable UPX by default
Disable UPX by default, even if its executable is available and use
of UPX is supported on the target OS (i.e., Windows).

Interested users must now explicitly enable use of UPX via new
`--upx` command-line option, or by setting `upx=True` in the .spec
file (and live with potential fallout).

This applies only to newly-generated spec files (including spec
files that are generated on the fly). The `--noupx` option is kept
for backward compatibility, although it is now effectively a no-op.