tests: fix the extended importlib.resources.files() test for python < 3.12
Passing a module name to `importlib.resources.files()` seems to be
supported only under python >= 3.12 (stdlib) or equivalent
`importlib_resources` >= 5.12. Under earlier versions, it raises
`TypeError: '<name>' is not a package`.

So allow `TypeError` to be raised by the test, in order to handle
situations where we are using python < 3.12 and stdlib version
of `importlib.resources` (or `importlib_resources` < 5.12).