modulegraph: fix display of aliased nodes in generated reports
Fix display of `AliasNode` in generated reports (cross-reference
html, dot file, report to stdout). In the case of `AliasNode`, the
`identifier` field contains the identifier of the target node.

However, `graphident` field contains the actual node's identifier
(as it is used for bookeeping in the `altgraph` `ObjectGraph`
backend), so the reporting code should use that instead.