tests: add a test for package with mixed module_collection_mode
Add a test for package that is collected with mixed
module_collection_mode, i.e., that has one module collected as a
source .py file only.

This verifies that PyInstaller's frozen importer and python's
`_frozen_importlib_external.PathFinder` complement, and not exclude,
each other. This in turn allows us to have packages that have modules
in PYZ archive, while their binary extensions are collected as
individual files (same goes for some pure-python modules that we
may need to collect in source-only form).

The same test is implcitly performed by any other test that uses
a package with pure-python modules and binary extensions, but it
is better to also have an explicit test.