tests: add tests for run-time sys.path modification
Add tests for run-time sys.path modifications, typically with aim
of exposing some part of a package to the outside world as a
top-level package; for example, to expose a vendored package to the
outside world.

These tests aim to verify that we can handle dynamic sys.path
modifications within PYZ-collected packages, while honoring the
order of entries in sys.path.