loader: rework PyiFrozenImporter into path entry finder implementation
This commit transforms our monolithic meta finder, PyiFrozenImporter,
into a path-bound path entry finder (`importlib.abc.PathEntryFinder`),
which is registered and used with python's own meta finder, `PathFinder`,
via the latter's extension hook mechanism (`sys.path_hooks`).

The monolithic finder/importer is impractical when it comes to supporting
dynamic `sys.path` modifications. The PEP302 loader extension methods
receive only fully-qualified module name (e.g., `mypackage.mymodule`) as
an argument, and a loader lacking path context will have difficult time
to decide whether that should, at PYZ archive level, translate into
`mypackage.mymodule` (for `sys._MEIPASS` in `sys.path`) or into
`myotherpackage._vendored.mypackage.mymodule` (for
`sys._MEIPASS/myotherpackage/_vendored` entry in `sys.path˙). Using
path-bound loaders/importers (similar to python's own `FileFinder`)
solves this issue.

When splitting the functionality into "top-level" meta finder and
path-bound loaders/importers, the design of python's `PathFinder` and
`FileFinder` was used as a reference. Therefore, it eventually became
clear that the new `PyiFrozenImporter` in fact became an implementation
of `importlib.abc.PathEntryFinder`, same as python's `FileFinder˙,
and could be used with python's `PathFinder`, instead of having its
own meta finder (whose main task became management of path-bound
importer instances).

The main problem with turning `PyiFrozenImporter` into a path entry
finder is that python's `PathFinder` allows only one path entry finder
instance per given path (i.e., the first one that is successfully
instantiated with the given path). Therefore, `PyiFrozenImporter`
and python's `FileFinder` cannot "cooperatively" handle the same path
at the `PathFinder` level. This is a problem, because we would like
`PyiFrozenImporter` to handle only modules in the PYZ archive (while
simulating on-filesystem paths for compatibility reasons), while actual
on-filesystem modules (binary extension modules, modules that are
collected only as source .py files, etc.) need to be handled by python's
`FileFinder` and whatever loader is applicable (`ExtensionFileLoader`,
`SourceFileLoader`).

To work around this limitation, we have the `PyiFrozenImporter` create
and manage a "fallback" instance of a `FileFinder`, and use it to
resolve the corresponding modules. For example, if we cannot return
a spec for a module based on a PYZ entry, we forward the call to
`find_spec` on the managed `FileFinder` instance, in case the module
is an extension or a source-only module. This way, we simulate
the two finders/importers operating in "complementary" mode (while
avoiding having to subclass the `FileFinder`).

The `PyiFrozenImporter` allows instantiation only on paths that are
rooted in `sys._MEIPASS` (or rather, on paths that are, once fully
resolved, rooted in fully-resolved `sys._MEIPASS`); this way, locations
completely external to the frozen application continue to be handled by
`FileFinder` only.

Similarly, PEP420 namespace packages with portions both inside the
frozen application (with PYZ-based portions and/or filesystem-only
portions) and in external locations should now be properly handled,
with all portions detected as parts of the namespace package.

Other changes to pyimod02_importers module:

Introduce a new minimal loader, `PyiFrozenEntryPointLoader`, that
allows the code object `__main__` module to be retrieved via `get_code()`
method.

The PYZ archive is now opened in `install()` function and stored as
a module-global variable.

The PYZ TOC tree is now stored as a module-global variable, and can
be accessed via `get_pyz_toc_tree()` function (first call computes
the tree).

Changes to related run-time hooks:

In `pkg_resources` run-time hook, we now need to register our
`PyiFrozenImporter` as both a loader and as a finder. Once the registration
is complete, we also need to force re-initialization of the master
working set, because original set is built with python's `FileFinder`
(still) handling the `sys._MEIPASS`.

The `pkgutil` run-time hook is greatly simplified, as we do not need to
override `pkgutil.iter_modules` implementation anymore. Instead, we can
provide an implementation of `iter_importer_modules` for our
`PyiFrozenImporter`. This implementation lists modules based on the
PYZ TOC tree, and also forwards the `iter_importer_modules` call to
the managed instance of `FileFinder`, so that on-filesystem modules are
also returned.