bootloader: support execution by direct dynamic loader invocation in onedir mode
Onedir mode on POSIX platform performs additional exec to correctly
set environment for itself. When binary is invoked as `ld-linux
binary.py` and the exec is not relying on provided dynamic loader but
relies instead on the one written in ELF header.

We already detect this way of invocation but we didn't kept this
information for the exec. This patch changes that.