bootloader: POSIX: if execvp() fails, exit immediately via exit()
If we fail to start the onefile child process via `execvp()` for
whatever reason, immediately exit the (forked) child process via
the `exit()` call. Otherwise, the child process will continue
executing the cleanup path, which is intended for the *parent*
process, and will end up interfering with the cleanup in the
actual parent process. For example, because it will remove the
temporary application directory, the parent process will fail
to do so in during its own cleanup.