isolated: implicitly disallow nesting of isolated subprocesses
When trying to use `PyInstaller.isolated` from a process that is
already a PyInstaller-isolated subprocess, make the `isolated`
framework no-op and execute the code directly in that process.

This means that the maximum depth of PyInstaller-isolated subprocess
tree is one, and that nesting of isolated subprocesses is implicitly
disallowed now.

This behavior enables transparent use of hook utility functions
within isolated subprocess, without having to worry whether their
implementation makes use of `isolated` framework or not. For example,
we can now import an expensive package once, and then run several
hook utility functions that perform their tasks in the same process,
instead of each spawning their own subprocess and importing the
package in it.