analysis: do not associate standard and run-time hooks with AliasNode
Remove the `AliasNode` from `PyInstaller.compat.VALID_MODULE_TYPES` set,
in order to prevent the standard hook for aliased module name from being
run. If such module needs to be hooked, the hook should be added for
its "true" name in order to properly accommodate the situation that
led to aliasing.

For example, if we end up adding alias that maps `mydependency` to
`mypackage._vendor.mydependency`, we should not run the hook for
`mydependency`, which would attempt to collect data files and binaries
for the "top-level" package (into corresponding "top-level" directory).
The collection of extra files for the vendored package should be handled
by the hook for vendored package, or hook for parent package that does
the vendoring.

In practice, things already kind of work this way, although it is
hard to say whether by design or due to an oversight. Although hook
for `mydependency` would run and register its `datas` and `binaries`
lists with `AdditionalFilesCache` under `mydependency` name, when the
time comes to gather these lists in `PyiModuleGraph.make_hook_binaries_toc`
and `make_hook_datas_toc`, the module name is taken as `node.identifier`,
which in `AliasNode` contains the name of referred module. So this ends
up trying to access binaries/datas for `mypackage._vendor.mydependency`,
twice.

So we are primarily aiming to formalize the behavior, but also to reduce
the noise during the build, and to avoid potential issues/discrepancies
with `hiddenimports`, especially if we have both "stand-alone" and
"vendored" hook available.

The situation with run-time hooks is quite similar to the above; the
`PyiModuleGraph.analyze_runtime_hooks` creates the list of modules based
on `VALID_MODULE_TYPES`, but `_make_toc` and its underlying `_node_to_toc`
again rely on `node.identifier` for the module name. So in practice,
run-time hook for `mydependency` (if available) is ignored, and the
run-time hook for `mypackage._vendor.mydependency` (if available) is
added twice, as evident by two "Including run-time hook..." messages
in the build log. However, the actual duplication is prevented by TOC
de-duplication mechanism that we have in place. So again, we are
primarily aiming to reduce noise and confusion.