hookutils: make copy_metadata aware of setuptools-vendored packages
Have `copy_metadata` check whether the metadata is located in
`setuptools/_vendor` directory; if it is, have the destination
use the same prefix, instead of collecting into top-level
application directory.

This way, we prevent potential duplication of metadata from
setuptools-vendored packages in the top-level application
directory.

With `copy_metadata` made aware of setuptools-vendored packages,
the same (potential) duplication problem is also resolved in
`collect_all`, `collect_entry_point`, as well as in the built-in
automatic metadata use detection and collection mechanism.