analysis: clean up the hooks cache and their loading
Our analysis system allows only one hook instance per module (to
avoid conflicts  between them), so instead of having
`ModuleHookCache` store lists of hooks per module, have it store
only the active/selected hook instance for each hooked module.

This ensures that the adjacent code that uses hook cache clearly
conveys that only one hook per module is allowed; previously, the
said code would loop over the list of hooks, and:
- in case of multiple standard hooks per module, the subsequent
  ones had been marked as "shallow" (inactive) when their `ModuleHook`
  instances were constructed, and were thus no-op.
- in case of pre-safe-import-module and pre-find-module-path hooks,
  all hook instances were (erroneously) ran.

The new approach ensures that selection is applied to all three
types of hooks, and that same selection mechanism is used; at the
moment, this is still the order in which hook directories are
processed, but we can later extend it with explicit hook priority
values.