analysis: use plain constants for default hook priority values
Forego the use of `enum.IntEnum` and define hook priority values
as "plain" constants instead. In addition to issues with marshaling
in isolated sub-process communication, the representation of enum
values is also incompatible with our repr-based serialization of
Target objects and their eval-based de-serialization, and thus
breaks caching (as well as couple of tests that try to read
generated .toc files).